+++ roman/roman_test.go
@@ -6,7 +6,10 @@ package roman
 
 import (
 	"errors"
+	"fmt"
 	"testing"
+
+	"go.lstv.dev/util/test"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -19,35 +22,59 @@ func Test_Number_MarshalText(t *testing.T) {
 		assert.Equal(t, DefaultFormat, f)
 		return []byte(`abc`), nil
 	}
-	b, err := Number(15749).MarshalText()
-	assert.Equal(t, []byte(`abc`), b)
-	assert.NoError(t, err)
-	formatError := errors.New("format error")
+	test.MarshalText(t, []test.CaseText[Number]{
+		{
+			Data:  `abc`,
+			Value: Number(15749),
+		},
+	})
+
 	Formatter = func(buf []byte, n Number, f Format) ([]byte, error) {
-		return nil, formatError
+		return nil, errors.New("format error")
 	}
-	b, err = Number(15749).MarshalText()
-	assert.Nil(t, b)
-	assert.Equal(t, formatError, err)
+	test.MarshalText(t, []test.CaseText[Number]{
+		{
+			Error: test.Error("format error"),
+			Value: Number(15749),
+		},
+	})
 }
 
 func Test_Number_UnmarshalText(t *testing.T) {
-	err := errors.New("error")
 	Parser = func(input []byte, r Rule) (Number, error) {
 		assert.Equal(t, []byte(`abc`), input)
 		assert.Equal(t, Rule(0), r)
-		return 0, err
+		return 0, errors.New("error")
 	}
-	n := Number(123)
-	assert.Equal(t, err, n.UnmarshalText([]byte(`abc`)))
-	assert.Equal(t, Number(123), n)
+	test.UnmarshalText(t, []test.CaseText[Number]{
+		{
+			Error: test.Error("error"),
+			Data:  `abc`,
+		},
+	}, nil)
+
 	Parser = func(input []byte, r Rule) (Number, error) {
 		assert.Equal(t, []byte(`abc`), input)
 		assert.Equal(t, Rule(0), r)
 		return Number(15749), nil
 	}
-	assert.NoError(t, n.UnmarshalText([]byte(`abc`)))
-	assert.Equal(t, Number(15749), n)
+	test.UnmarshalText(t, []test.CaseText[Number]{
+		{
+			Data:  `abc`,
+			Value: Number(15749),
+		},
+	}, nil)
+}
+
+func Test_Number_Format(t *testing.T) {
+	Formatter = DefaultFormatter
+	n := Number(4)
+	assert.Equal(t, `roman.Number`, fmt.Sprintf("%T", n))
+	assert.Equal(t, `IIII`, fmt.Sprintf("%L", n))
+	assert.Equal(t, `iiii`, fmt.Sprintf("%l", n))
+	assert.Equal(t, `IV`, fmt.Sprintf("%R", n))
+	assert.Equal(t, `iv`, fmt.Sprintf("%r", n))
+	assert.Equal(t, `IV`, fmt.Sprintf("%s", n))
 }
 
 func Test_Number_String(t *testing.T) {
@@ -59,9 +86,35 @@ func Test_Number_String(t *testing.T) {
 		return []byte(`abc`), nil
 	}
 	assert.Equal(t, `abc`, Number(15749).String())
-	formatError := errors.New("format error")
+
 	Formatter = func(buf []byte, n Number, f Format) ([]byte, error) {
-		return nil, formatError
+		return nil, errors.New("format error")
 	}
 	assert.Equal(t, `MMMMMMMMMMMMMMMDCCXXXXIX`, Number(15749).String())
+}
+
+func Test_Number_format(t *testing.T) {
+	DefaultFormat = 0
+	Formatter = func(buf []byte, n Number, f Format) ([]byte, error) {
+		assert.Nil(t, buf)
+		assert.Equal(t, Number(15749), n)
+		assert.Equal(t, FormatLong40, f)
+		return []byte(`abc`), nil
+	}
+	assert.Equal(t, []byte(`abc`), Number(15749).format(FormatLong40))
+
+	Formatter = func(buf []byte, n Number, f Format) ([]byte, error) {
+		return nil, errors.New("format error")
+	}
+	assert.Equal(t, `MMMMMMMMMMMMMMMDCCXLIX`, Number(15749).String())
+}
+
+func Test_formatByVerb(t *testing.T) {
+	DefaultFormat = FormatLong40
+	assert.Equal(t, DefaultFormat, formatByVerb(' '))
+	assert.Equal(t, DefaultFormat, formatByVerb('s'))
+	assert.Equal(t, FormatLowerCase, formatByVerb('r'))
+	assert.Equal(t, Format(0), formatByVerb('R'))
+	assert.Equal(t, FormatLowerCase|FormatLong, formatByVerb('l'))
+	assert.Equal(t, FormatLong, formatByVerb('L'))
 }