+++ sem/version_test.go
@@ -6,8 +6,11 @@ package sem
 
 import (
 	"errors"
+	"fmt"
 	"math"
 	"testing"
+
+	"go.lstv.dev/util/test"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -342,6 +345,63 @@ func Test_Ver_Valid(t *testing.T) {
 	assert.Error(t, v.Valid())
 }
 
+func Test_Ver_IsZero(t *testing.T) {
+	v := Ver{}
+	assert.True(t, v.IsZero())
+	v = Ver{Major: 1}
+	assert.False(t, v.IsZero())
+	v = Ver{Minor: 1}
+	assert.False(t, v.IsZero())
+	v = Ver{Patch: 1}
+	assert.False(t, v.IsZero())
+	v = Ver{Build: "a"}
+	assert.False(t, v.IsZero())
+	v = Ver{PreRelease: "b"}
+	assert.False(t, v.IsZero())
+}
+
+func Test_Ver_Core(t *testing.T) {
+	v := Ver{
+		Major:      1,
+		Minor:      2,
+		Patch:      3,
+		PreRelease: "",
+		Build:      "",
+	}
+	assert.Equal(t, Ver{
+		Major:      1,
+		Minor:      2,
+		Patch:      3,
+		PreRelease: "",
+		Build:      "",
+	}, v.Core())
+	v.Build = "a"
+	assert.Equal(t, Ver{
+		Major:      1,
+		Minor:      2,
+		Patch:      3,
+		PreRelease: "",
+		Build:      "",
+	}, v.Core())
+	v.Build = "a"
+	v.PreRelease = "b"
+	assert.Equal(t, Ver{
+		Major:      1,
+		Minor:      2,
+		Patch:      3,
+		PreRelease: "",
+		Build:      "",
+	}, v.Core())
+	v.Build = ""
+	assert.Equal(t, Ver{
+		Major:      1,
+		Minor:      2,
+		Patch:      3,
+		PreRelease: "",
+		Build:      "",
+	}, v.Core())
+}
+
 func Test_Ver_NextMajor(t *testing.T) {
 	assert.Equal(t, Ver{
 		Major:      1,
@@ -469,16 +529,19 @@ func Test_Ver_MarshalText(t *testing.T) {
 		assert.Equal(t, Format(0), f)
 		return []byte(`ab`), nil
 	}
-	b, err := Ver{
-		Major:      2,
-		Minor:      1,
-		Patch:      3,
-		PreRelease: "a",
-		Build:      "b",
-	}.MarshalText()
-	assert.Equal(t, []byte(`ab`), b)
-	assert.NoError(t, err)
-	formatError := errors.New("format error")
+	test.MarshalText(t, []test.CaseText[Ver]{
+		{
+			Data: `ab`,
+			Value: Ver{
+				Major:      2,
+				Minor:      1,
+				Patch:      3,
+				PreRelease: "a",
+				Build:      "b",
+			},
+		},
+	})
+
 	Formatter = func(buf []byte, v Ver, f Format) ([]byte, error) {
 		assert.Nil(t, buf)
 		assert.Equal(t, Ver{
@@ -489,17 +552,20 @@ func Test_Ver_MarshalText(t *testing.T) {
 			Build:      "b",
 		}, v)
 		assert.Equal(t, Format(0), f)
-		return nil, formatError
+		return nil, errors.New("format error")
 	}
-	b, err = Ver{
-		Major:      2,
-		Minor:      1,
-		Patch:      3,
-		PreRelease: "a",
-		Build:      "b",
-	}.MarshalText()
-	assert.Nil(t, b)
-	assert.Equal(t, formatError, err)
+	test.MarshalText(t, []test.CaseText[Ver]{
+		{
+			Error: test.Error("format error"),
+			Value: Ver{
+				Major:      2,
+				Minor:      1,
+				Patch:      3,
+				PreRelease: "a",
+				Build:      "b",
+			},
+		},
+	})
 }
 
 func Test_Ver_UnmarshalText(t *testing.T) {
@@ -514,29 +580,37 @@ func Test_Ver_UnmarshalText(t *testing.T) {
 			Build:      "b",
 		}, nil
 	}
-	v := Ver{}
-	assert.NoError(t, v.UnmarshalText([]byte(`ab`)))
-	assert.Equal(t, Ver{
-		Major:      1,
-		Minor:      2,
-		Patch:      3,
-		PreRelease: "a",
-		Build:      "b",
-	}, v)
-	parseError := errors.New("parse error")
+	test.UnmarshalText(t, []test.CaseText[Ver]{
+		{
+			Data: `ab`,
+			Value: Ver{
+				Major:      1,
+				Minor:      2,
+				Patch:      3,
+				PreRelease: "a",
+				Build:      "b",
+			},
+		},
+	}, nil)
+
 	Parser = func(input []byte, r Rule) (v Ver, err error) {
 		assert.Equal(t, []byte(`ab`), input)
 		assert.Equal(t, Rule(0), r)
-		return Ver{}, parseError
+		return Ver{}, errors.New("parse error")
 	}
-	assert.Equal(t, parseError, v.UnmarshalText([]byte(`ab`)))
-	assert.Equal(t, Ver{
-		Major:      1,
-		Minor:      2,
-		Patch:      3,
-		PreRelease: "a",
-		Build:      "b",
-	}, v)
+	test.UnmarshalText(t, []test.CaseText[Ver]{
+		{
+			Error: test.Error("parse error"),
+			Data:  `ab`,
+		},
+	}, nil)
+}
+
+func Test_Ver_Format(t *testing.T) {
+	Formatter = DefaultFormatter
+	assert.Equal(t, "sem.Ver", fmt.Sprintf("%T", New(1, 2, 3, "b", "p")))
+	assert.Equal(t, "v1.2.3-b+p", fmt.Sprintf("%t", New(1, 2, 3, "b", "p")))
+	assert.Equal(t, "1.2.3-b+p", fmt.Sprintf("%s", New(1, 2, 3, "b", "p")))
 }
 
 func Test_Ver_StringTag(t *testing.T) {
@@ -597,7 +671,7 @@ func Test_Ver_String(t *testing.T) {
 		Build:      "b",
 	}
 	assert.Equal(t, `x`, v.String())
-	formatError := errors.New("format error")
+
 	Formatter = func(buf []byte, v Ver, f Format) ([]byte, error) {
 		assert.Nil(t, buf)
 		assert.Equal(t, Ver{
@@ -608,7 +682,50 @@ func Test_Ver_String(t *testing.T) {
 			Build:      "b",
 		}, v)
 		assert.Equal(t, Format(0), f)
-		return nil, formatError
+		return nil, errors.New("format error")
 	}
 	assert.Equal(t, `1.2.3-a+b`, v.String())
+}
+
+func Test_Ver_format(t *testing.T) {
+	Formatter = func(buf []byte, v Ver, f Format) ([]byte, error) {
+		assert.Nil(t, buf)
+		assert.Equal(t, Ver{
+			Major:      1,
+			Minor:      2,
+			Patch:      3,
+			PreRelease: "a",
+			Build:      "b",
+		}, v)
+		assert.Equal(t, FormatTag, f)
+		return []byte(`x`), nil
+	}
+	v := Ver{
+		Major:      1,
+		Minor:      2,
+		Patch:      3,
+		PreRelease: "a",
+		Build:      "b",
+	}
+	assert.Equal(t, []byte(`x`), v.format(FormatTag))
+
+	Formatter = func(buf []byte, v Ver, f Format) ([]byte, error) {
+		assert.Nil(t, buf)
+		assert.Equal(t, Ver{
+			Major:      1,
+			Minor:      2,
+			Patch:      3,
+			PreRelease: "a",
+			Build:      "b",
+		}, v)
+		assert.Equal(t, FormatTag, f)
+		return nil, errors.New("format error")
+	}
+	assert.Equal(t, []byte(`v1.2.3-a+b`), v.format(FormatTag))
+}
+
+func Test_formatByVerb(t *testing.T) {
+	assert.Equal(t, Format(0), formatByVerb(' '))
+	assert.Equal(t, Format(0), formatByVerb('s'))
+	assert.Equal(t, FormatTag, formatByVerb('t'))
 }