+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ## [Unreleased]
 
+## [0.5.0] - 2022-03-23
+### Added
+- Format `roman.FormatLowerCase`.
+- Methods `sem.Ver.Core` and `sem.Ver.IsZero`.
+- Implementation of `fmt.Formatter` for:
+  - `date.Date`
+  - `roman.Number`
+  - `sem.Ver`
+- Package `test` for enhanced testing.
+
+### Changed
+- Implementation of tests.
+
+### Fixed
+- Method `Date.MarshalBinary` error.
+
 ## [0.4.0] - 2022-03-16
 ### Added
 - Package `constraint` with generic support and helpers.
@@ -72,7 +88,8 @@
 ### Added
 - First release of util.
 
-[Unreleased]: https://github.com/livesport-tv/util/compare/v0.4.0...master
+[Unreleased]: https://github.com/livesport-tv/util/compare/v0.5.0...master
+[0.5.0]: https://github.com/livesport-tv/util/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/livesport-tv/util/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/livesport-tv/util/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/livesport-tv/util/compare/v0.2.0...v0.2.1